+++ CHANGELOG.md
@@ -11,14 +11,31 @@
 
 ### Devop
 
--   refactor Home View for i18n [#1142](https://github.com/EthVM/EthVM/pull/1142)
--   Added Portfolio View for i18n [#1155](https://github.com/EthVM/EthVM/pull/1155)
--   Added setting View for i18n [#1158](https://github.com/EthVM/EthVM/pull/1158)
--   Added Transaction details View for i18n [#1159](https://github.com/EthVM/EthVM/pull/1159)
+-  refactor Home View for i18n [#1142](https://github.com/EthVM/EthVM/pull/1142)
+-  added Portfolio View for i18n [#1155](https://github.com/EthVM/EthVM/pull/1155)
+-  added setting View for i18n [#1158](https://github.com/EthVM/EthVM/pull/1158)
+-  added Transaction details View for i18n [#1159](https://github.com/EthVM/EthVM/pull/1159)
 
 ### Feature
 
--   # advertise with us page [#1156](https://github.com/EthVM/EthVM/pull/1156)
+- advertise with us page [#1156](https://github.com/EthVM/EthVM/pull/1156)
+- implemented GA [#1160](https://github.com/EthVM/EthVM/pull/1160)
+- add new privacy policy route [#1160](https://github.com/EthVM/EthVM/pull/1160)
+
+
+### UI
+- add small and large promo buttons [#1160](https://github.com/EthVM/EthVM/pull/1160)
+- new address view layout [#1160](https://github.com/EthVM/EthVM/pull/1160)
+
+======================================================================
+
+### Release v2.1.1-hotfix.1
+
+###  Fix
+- update Trasfers query name [#1163](https://github.com/EthVM/EthVM/pull/1163)
+
+======================================================================
+
 
 ### Release v2.1.1